screen-orientation/onchange-event.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(t => {
  const type = screen.orientation.type;
  screen.orientation.addEventListener('change', t.unreached_func("change event should not be fired"));
  return screen.orientation.lock(type).then(() => {
    assert_equals(screen.orientation.type, type);
  });
}, "Test that orientationchange event is not fired when the orientation does not change.");

promise_test(async t => {
  const orientations = [
    'portrait-primary',
    'portrait-secondary',
    'landscape-primary',
    'landscape-secondary'
  ];
  const orientationWatcher = new EventWatcher(t, screen.orientation, 'change');

  let currentIndex = orientations.indexOf(screen.orientation.type);
  let nextIndex;
  for (let i = 0; i < orientations.length; i++) {
    nextIndex = (currentIndex + 1) % orientations.length;
    screen.orientation.lock(orientations[nextIndex]);
    currentIndex = nextIndex;
    await orientationWatcher.wait_for('change');
    assert_equals(screen.orientation.type, orientations[currentIndex]);
  }
  screen.orientation.unlock();
}, "Test that orientationchange event is fired when the orientation changes.");
</script>
